+++ Localization/enUS.lua
@@ -237,6 +237,9 @@ L["Spider Soul Guide"] = "Go through a Rift Portal using a %s. There are 4 cache
 L["Sporebat Soul"] = true
 L["Sporebat Soul Guide"] = "Can drop from the last boss of any Burning Crusade Timewalking dungeon."
 
+L["Squirrel Soul"] = true
+L["Squirrel Soul Guide"] = "The first crittershape of your collection, courtesy of Choofa!"
+
 L["Stag Soul"] = true
 L["Stag Soul Guide"] = "Can be randomly offered as a reward from any Night Fae daily in Korthia."