+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added a filter to filter soulshapes by the patch they became available to players
 * Updated Russian translation (thanks to ZamestoTV)
+* Added the default Squirrel Crittershape (thanks to covsoul) 
 
 # 0.9.1